Shark + Fireball, Player + Seaweed Collisions
Code collisions with sharks and fireballs so that sharks disappear when in contact with the fireball and then the fireball is destroyed. The player can only shoot 5 fireballs at a time. When player collides with seaweed the speed decreases for 2 seconds and then returns to the normal speed. Error: Distance change rate stays the same, will have to change.