Fixed Different Distance Error
The Distance would reach zero too early and the game would end when the player was still in the middle of the screen. Now that doesn't happen, however now the speed of the player has changed and the game therefore ends a little earlier but there doesn't seem to be a huge difference in the playing experience.